fix: merged backend and api directories
