+++ src/main/scala/api/domain/Errors.scala
@@ -1,4 +1,4 @@
-package backend
+package api.domain
 
 sealed trait ErrorResponse:
   val message: String
