+++ src/main/scala/api/domain/Ingredient.scala
@@ -2,6 +2,8 @@ package api.domain
 
 import api.domain.IngredientId
 
+type IngredientId = BaseId
+
 case class Ingredient(
                      ingredientId: IngredientId,
                      name: String
