+++ src/main/scala/api/domain/Storage.scala
@@ -1,8 +1,10 @@
 package api.domain
 
+type StorageId = BaseId
+
 case class Storage(
                   storageId: StorageId,
                   ownerId: UserId,
-                  members: List[UserId], 
+                  members: List[UserId],
                   ingredients: List[IngredientId]
                   )
