+++ src/main/scala/api/domain/TypeAliases.scala
@@ -1,6 +1,3 @@
 package api.domain
 
 type BaseId = Long
-type IngredientId = BaseId
-type UserId = BaseId
-type StorageId = BaseId
