+++ src/main/scala/api/domain/User.scala
@@ -1,5 +1,7 @@
 package api.domain
 
+type UserId = BaseId
+
 case class User(
                userId: UserId,
                username: String,
