--- src/main/scala/backend/DomainModels.scala
@@ -1,6 +1,0 @@
-package backend
-
-type IngredientId = Int
-type StorageId = Int
-
-case class Ingredient(id: IngredientId, name: String)
